Implement sequelize with mysql and sqlite + fixes
- move global types to backend
- Fix precheck + test
- Fix for setting prod port
