+++ .env.example
@@ -1,9 +1,18 @@
 #  What env variables are required to run the application on production (comma separated values)
-PRECHECK_REQUIRED_ENV_VARS=APP_HOST_URL
+PRECHECK_REQUIRED_ENV_VARS=APP_HOST_URL,APP_DB_URL
 
 # Host of the application eg. https://somedomain.com or http://localhost:5000
 APP_HOST_URL=
 
 # The port on which the production server will start
-# it is useful if you plan to run the application directly (without using e.g. a docker)
+# it is useful if you plan to run the application directly (without using eg. a docker)
 APP_PROD_PORT=
+
+# Connection URL to database. Currently there are 2 available dialects:
+#
+# * SQLite 3  (URL format: "sqlite://FILE_NAME")
+#   eg.       sqlite://database.sqlite3
+# 
+# * MySQL     (URL format: "mysql://USER:PASS@HOST:PORT/DB_NAME")
+#   eg.       mysql://root:root@localhost/boilerplate_demo
+APP_DB_URL=
