+++ .gitignore
@@ -17,3 +17,4 @@ yarn-debug.log*
 yarn-error.log*
 
 # *** Backend ***
+*.sqlite3
