+++ backend/.eslintrc.json
@@ -31,6 +31,7 @@
     "semi": [
       "error",
       "always"
-    ]
+    ],
+    "@typescript-eslint/no-namespace": "off"
   }
 }
