+++ backend/package.json
@@ -13,7 +13,10 @@
     "axios": "^0.25.0",
     "express": "^4.17.2",
     "express-graphql": "^0.12.0",
-    "graphql": "^16.3.0"
+    "graphql": "^16.3.0",
+    "mysql2": "^2.3.3",
+    "sequelize": "^6.16.0",
+    "sqlite3": "^5.0.2"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
