+++ backend/src/boot/server.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import express from 'express';
-import { getConfig } from './config';
-import routes from './routes';
+import { getConfig } from '../config';
+import routes from '../routes';
 
 const runServer = (): Promise<express.Application> => new Promise((resolve) => {
   const { host, port } = getConfig();
@@ -18,11 +18,9 @@ const runServer = (): Promise<express.Application> => new Promise((resolve) => {
   // Run server
   app.listen(port, () => {
     const mode = process.env.NODE_ENV?.toUpperCase();
-    console.log(`Server is running in ${mode} mode on ${host}`);
+    console.log(`✅ Server is running in ${mode} mode on ${host}`);
     resolve(app);
   });
 });
 
-export {
-  runServer
-};
+export { runServer };
