+++ backend/tsconfig.build.json
@@ -2,6 +2,6 @@
   "extends": "./tsconfig.json",
   "exclude": [
     "**/*.test.ts",
-    "../@types/**/*.ts"
+    "@types/**/*.ts"
   ]
 }
