+++ backend/tsconfig.json
@@ -16,6 +16,6 @@
   ],
   "include": [
     "src/**/*.ts",
-    "../@types/**/*.ts"
+    "@types/**/*.ts"
   ]
 }
